Unify help menu tests in a single file
