+++ napari/_qt/_qapp_model/_tests/test_help_menu_urls.py
@@ -1,8 +1,12 @@
-"""For testing the URLs in the Help menu"""
+"""For testing the Help menu"""
+
+import sys
+from unittest import mock
 
 import pytest
 import requests
 
+from napari._app_model import get_app
 from napari._qt._qapp_model.qactions._help import HELP_URLS
 
 
@@ -13,3 +17,23 @@ def test_help_urls(url):
 
     r = requests.head(HELP_URLS[url])
     r.raise_for_status()
+
+
+@pytest.mark.parametrize(
+    'action_id',
+    [
+        'napari.window.help.info',
+        'napari.window.help.about_macos',
+    ]
+    if sys.platform == 'darwin'
+    else ['napari.window.help.info'],
+)
+def test_about_action(make_napari_viewer, action_id):
+    app = get_app()
+    viewer = make_napari_viewer()
+
+    with mock.patch(
+        'napari._qt.dialogs.qt_about.QtAbout.showAbout'
+    ) as mock_about:
+        app.commands.execute_command(action_id)
+    mock_about.assert_called_once_with(viewer.window._qt_window)
